requirements.txt
matplotlib==3.5.3
pandas==1.4.4
plotly==5.11.0
pyairtable==1.3.0
python-decouple==3.6
streamlit==1.11.0
